+++ src/main/java/nextstep/auth/AuthService.java
@@ -1,10 +1,10 @@
 package nextstep.auth;
 
 import nextstep.auth.role.Role;
-import nextstep.exception.BusinessErrorCode;
-import nextstep.exception.BusinessException;
-import nextstep.exception.DataAccessErrorCode;
-import nextstep.exception.DataAccessException;
+import nextstep.exception.business.BusinessErrorCode;
+import nextstep.exception.business.BusinessException;
+import nextstep.exception.dataaccess.DataAccessErrorCode;
+import nextstep.exception.dataaccess.DataAccessException;
 import nextstep.member.Member;
 import nextstep.member.MemberDao;
 import org.springframework.stereotype.Service;
