+++ src/main/java/nextstep/member/MemberService.java
@@ -1,9 +1,9 @@
 package nextstep.member;
 
-import nextstep.exception.BusinessErrorCode;
-import nextstep.exception.BusinessException;
-import nextstep.exception.DataAccessErrorCode;
-import nextstep.exception.DataAccessException;
+import nextstep.exception.business.BusinessErrorCode;
+import nextstep.exception.business.BusinessException;
+import nextstep.exception.dataaccess.DataAccessErrorCode;
+import nextstep.exception.dataaccess.DataAccessException;
 import org.springframework.stereotype.Service;
 
 @Service
